backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
google-genai>=1.0.0
playwright==1.48.0
Pillow==11.0.0
typing-extensions>=4.8.0